code/src/backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
werkzeug==2.3.7
python-dotenv==1.0.0
pdfplumber==0.10.2
pytesseract==0.3.10
pillow==10.0.1
openai==0.28.1
sentence-transformers==2.2.2
scikit-learn==1.3.0